+++ src/aui-data-editor/js/aui-options-data-editor.js
@@ -26,7 +26,7 @@ var CSS_EDITOR = A.getClassName('options', 'data', 'editor'),
 A.OptionsDataEditor = A.Base.create('options-data-editor', A.DataEditor, [], {
     TPL_EDITOR_CONTENT: '<div class="' + CSS_EDITOR + '">' +
         '<div class="' + CSS_EDITOR_OPTIONS + '"></div>' +
-        '<button class="' + CSS_EDITOR_ADD + '">Tap to add an option</button></div>',
+        '<button class="' + CSS_EDITOR_ADD + '">{addOption}</button></div>',
     TPL_EDITOR_OPTION: '<div class="' + CSS_EDITOR_OPTION + '">' +
         '<span class="' + CSS_EDITOR_OPTION_HANDLE + ' glyphicon glyphicon-sort"></span>' +
         '<input class="' + CSS_EDITOR_OPTION_TEXT + ' type="text"" value="{text}"></input>' +
@@ -334,6 +334,19 @@ A.OptionsDataEditor = A.Base.create('options-data-editor', A.DataEditor, [], {
          */
         originalValue: {
             value: []
+        },
+
+        /**
+         * Collection of strings used to label elements of the UI.
+         *
+         * @attribute strings
+         * @type {Object}
+         */
+        strings: {
+            value: {
+                addOption: 'Tap to add an option',
+                required: 'REQUIRED'
+            }
         }
     }
 });