+++ src/aui-layout/js/aui-layout-builder.js
@@ -146,6 +146,20 @@ A.LayoutBuilder = A.Base.create('layout-builder', A.Base, [
             valueFn: function() {
                 return new A.Layout();
             }
+        },
+
+        /**
+         * Collection of strings used to label elements of the UI.
+         *
+         * @attribute strings
+         * @type {Object}
+         */
+        strings: {
+            value: {
+                addColumn: 'Add Column',
+                removeRow: 'Remove Row'
+            },
+            writeOnce: true
         }
     }
 });